requirements.txt
streamlit
tensorflow
scikit-learn
pandas
numpy
mysql-connector-python
joblib